+++ apps/block_scout_web/assets/js/lib/history_chart.js
@@ -27,14 +27,6 @@ const grid = {
   drawOnChartArea: false
 }
 
-function getTxChartColor () {
-  if ((isDarkMode())) {
-    return sassVariables.dashboardLineColorTransactionsDarkTheme
-  } else {
-    return sassVariables.dashboardLineColorTransactions
-  }
-}
-
 function getPriceChartColor () {
   if ((isDarkMode())) {
     return sassVariables.dashboardLineColorPriceDarkTheme
@@ -257,8 +249,24 @@ class MarketHistoryChart {
       cubicInterpolationMode: 'monotone',
       fill: false,
       pointRadius: 0,
-      backgroundColor: getTxChartColor(),
-      borderColor: getTxChartColor()
+      backgroundColor: function (context) {
+        const chart = context.chart
+        const { ctx, chartArea } = chart
+
+        if (!chartArea) {
+          return
+        }
+        return getGradient(ctx, chartArea)
+      },
+      borderColor: function (context) {
+        const chart = context.chart
+        const { ctx, chartArea } = chart
+
+        if (!chartArea) {
+          return
+        }
+        return getGradient(ctx, chartArea)
+      }
       // lineTension: 0
     }
 
@@ -336,6 +344,14 @@ export function createMarketHistoryChart (el) {
       })
   })
   return chart
+}
+
+function getGradient (ctx, chartArea) {
+  const gradient = ctx.createLinearGradient(0, chartArea.bottom, 0, chartArea.top)
+  gradient.addColorStop(0, '#E93434')
+  gradient.addColorStop(1, '#861BF5')
+
+  return gradient
 }
 
 $('[data-chart-error-message]').on('click', _event => {