+++ apps/indexer/test/indexer/block/realtime/fetcher_test.exs
@@ -489,12 +489,6 @@ defmodule Indexer.Block.Realtime.FetcherTest do
               jsonrpc: "2.0",
               method: "eth_getBalance",
               params: ["0x66c9343c7e8ca673a1fedf9dbf2cd7936dbbf7e3", "0x3C3660"]
-            },
-            %{
-              id: 3,
-              jsonrpc: "2.0",
-              method: "eth_getBalance",
-              params: ["0x698bf6943bab687b2756394624aa183f434f65da", "0x3C365F"]
             }
           ],
           _ ->
@@ -502,8 +496,7 @@ defmodule Indexer.Block.Realtime.FetcherTest do
              [
                %{id: 0, jsonrpc: "2.0", result: "0x148adc763b603291685"},
                %{id: 1, jsonrpc: "2.0", result: "0x53474fa377a46000"},
-               %{id: 2, jsonrpc: "2.0", result: "0x53507afe51f28000"},
-               %{id: 3, jsonrpc: "2.0", result: "0x3e1a95d7517dc197108"}
+               %{id: 2, jsonrpc: "2.0", result: "0x53507afe51f28000"}
              ]}
         end)
       end
@@ -514,8 +507,7 @@ defmodule Indexer.Block.Realtime.FetcherTest do
                   addresses: [
                     %Address{hash: first_address_hash, fetched_coin_balance_block_number: 3_946_079},
                     %Address{hash: second_address_hash, fetched_coin_balance_block_number: 3_946_079},
-                    %Address{hash: third_address_hash, fetched_coin_balance_block_number: 3_946_080},
-                    %Address{hash: fourth_address_hash, fetched_coin_balance_block_number: 3_946_079}
+                    %Address{hash: third_address_hash, fetched_coin_balance_block_number: 3_946_080}
                   ],
                   address_coin_balances: [
                     %{
@@ -529,10 +521,6 @@ defmodule Indexer.Block.Realtime.FetcherTest do
                     %{
                       address_hash: third_address_hash,
                       block_number: 3_946_080
-                    },
-                    %{
-                      address_hash: fourth_address_hash,
-                      block_number: 3_946_079
                     }
                   ],
                   blocks: [%Chain.Block{number: 3_946_079}, %Chain.Block{number: 3_946_080}],
@@ -981,12 +969,6 @@ defmodule Indexer.Block.Realtime.FetcherTest do
               jsonrpc: "2.0",
               method: "eth_getBalance",
               params: ["0x66c9343c7e8ca673a1fedf9dbf2cd7936dbbf7e3", "0x3C3660"]
-            },
-            %{
-              id: 3,
-              jsonrpc: "2.0",
-              method: "eth_getBalance",
-              params: ["0x698bf6943bab687b2756394624aa183f434f65da", "0x3C365F"]
             }
           ],
           _ ->
@@ -994,8 +976,7 @@ defmodule Indexer.Block.Realtime.FetcherTest do
              [
                %{id: 0, jsonrpc: "2.0", result: "0x148adc763b603291685"},
                %{id: 1, jsonrpc: "2.0", result: "0x53474fa377a46000"},
-               %{id: 2, jsonrpc: "2.0", result: "0x53507afe51f28000"},
-               %{id: 3, jsonrpc: "2.0", result: "0x3e1a95d7517dc197108"}
+               %{id: 2, jsonrpc: "2.0", result: "0x53507afe51f28000"}
              ]}
         end)
       end
@@ -1009,8 +990,7 @@ defmodule Indexer.Block.Realtime.FetcherTest do
                   addresses: [
                     %Address{hash: first_address_hash, fetched_coin_balance_block_number: 3_946_079},
                     %Address{hash: second_address_hash, fetched_coin_balance_block_number: 3_946_079},
-                    %Address{hash: third_address_hash, fetched_coin_balance_block_number: 3_946_080},
-                    %Address{hash: fourth_address_hash, fetched_coin_balance_block_number: 3_946_079}
+                    %Address{hash: third_address_hash, fetched_coin_balance_block_number: 3_946_080}
                   ],
                   address_coin_balances: [
                     %{
@@ -1024,10 +1004,6 @@ defmodule Indexer.Block.Realtime.FetcherTest do
                     %{
                       address_hash: third_address_hash,
                       block_number: 3_946_080
-                    },
-                    %{
-                      address_hash: fourth_address_hash,
-                      block_number: 3_946_079
                     }
                   ],
                   block_rewards: [